rename selection tooltip converter
